Added base for new filesystem API.
This will be the new filesystem API, it is more like a *nix API.
It bases everything on nodes instead of a class hierarchy.

There is also a small example filesystem called NullFS, and a test
routine in kmain to test the new API (This one will be removed as soon
as the userspace gets access to this filesystem instead of the current
one).

Signed-off-by: Dan Printzell <xwildn00bx@gmail.com>